+++ demo/demo.component.ts
@@ -16,8 +16,15 @@ import { HighlightTag } from '../src/';
       <mwl-text-input-highlight
         [tags]="tags"
         [tagCssClass]="'bg-blue'"
-        [textInputElement]="textarea">
+        [textInputElement]="textarea"
+        (tagMouseEnter)="addDarkClass($event.target)"
+        (tagMouseLeave)="removeDarkClass($event.target)"
+        (tagClick)="tagClicked = $event.tag">
       </mwl-text-input-highlight>
+    </div>
+    <br>
+    <div class="alert alert-info" *ngIf="tagClicked">
+      Tag clicked! {{ tagClicked.data }}
     </div>
   `,
   styles: [
@@ -27,6 +34,12 @@ import { HighlightTag } from '../src/';
       }
       .bg-pink {
         background-color: lightcoral;
+      }
+      .bg-blue-dark {
+        background-color: #86c5da;
+      }
+      .bg-pink-dark {
+        background-color: #eb5252;
       }
       textarea {
         width: 500px;
@@ -39,6 +52,8 @@ export class DemoComponent implements OnInit {
   text: string = 'Hello @mattlewis92 how are you today?\n\nLook I have a #different background color!\n\n@angular is pretty awesome!';
 
   tags: HighlightTag[] = [];
+
+  tagClicked: HighlightTag;
 
   ngOnInit(): void {
     this.addTags();
@@ -54,7 +69,8 @@ export class DemoComponent implements OnInit {
         indices: {
           start: mention.index,
           end: mention.index + mention[1].length
-        }
+        },
+        data: mention[1]
       });
     }
 
@@ -67,8 +83,22 @@ export class DemoComponent implements OnInit {
           start: hashtag.index,
           end: hashtag.index + hashtag[1].length
         },
-        cssClass: 'bg-pink'
+        cssClass: 'bg-pink',
+        data: hashtag[1]
       });
     }
+  }
+
+  addDarkClass(elm: HTMLElement) {
+    if (elm.classList.contains('bg-blue')) {
+      elm.classList.add('bg-blue-dark');
+    } else if (elm.classList.contains('bg-pink')) {
+      elm.classList.add('bg-pink-dark');
+    }
+  }
+
+  removeDarkClass(elm: HTMLElement) {
+    elm.classList.remove('bg-blue-dark');
+    elm.classList.remove('bg-pink-dark');
   }
 }