+++ src/text-input-highlight.component.ts
@@ -1,9 +1,12 @@
 import {
+  ChangeDetectorRef,
   Component,
   ElementRef,
+  EventEmitter,
   Input,
   OnChanges,
   OnDestroy,
+  Output,
   Renderer2,
   SimpleChanges,
   ViewChild
@@ -51,6 +54,8 @@ const styleProperties = Object.freeze([
   'MozTabSize'
 ]);
 
+const tagIndexIdPrefix = 'text-highlight-tag-id-';
+
 function indexIsInsideTag(index: number, tag: HighlightTag) {
   return tag.indices.start < index && index < tag.indices.end;
 }
@@ -60,6 +65,19 @@ function overlaps(tagA: HighlightTag, tagB: HighlightTag) {
     indexIsInsideTag(tagB.indices.start, tagA) ||
     indexIsInsideTag(tagB.indices.end, tagA)
   );
+}
+
+function isCoordinateWithinRect(rect: ClientRect, x: number, y: number) {
+  return rect.left < x && x < rect.right && (rect.top < y && y < rect.bottom);
+}
+
+function escapeHtml(str: string): string {
+  return str.replace(/</g, '&lt;').replace(/>/g, '&gt;');
+}
+
+export interface TagMouseEvent {
+  tag: HighlightTag;
+  target: HTMLElement;
 }
 
 @Component({
@@ -85,9 +103,24 @@ export class TextInputHighlightComponent implements OnChanges, OnDestroy {
   @Input() tags: HighlightTag[] = [];
 
   /**
-   * The textarea to highlight. Currently only works with textareas.
+   * The textarea to highlight
    */
   @Input() textInputElement: HTMLTextAreaElement;
+
+  /**
+   * Called when the area over a tag is clicked
+   */
+  @Output() tagClick = new EventEmitter<TagMouseEvent>();
+
+  /**
+   * Called when the area over a tag is moused over
+   */
+  @Output() tagMouseEnter = new EventEmitter<TagMouseEvent>();
+
+  /**
+   * Called when the area over the tag has the mouse is removed from it
+   */
+  @Output() tagMouseLeave = new EventEmitter<TagMouseEvent>();
 
   /**
    * @private
@@ -103,7 +136,14 @@ export class TextInputHighlightComponent implements OnChanges, OnDestroy {
 
   private textareaEventListeners: Array<() => void> = [];
 
-  constructor(private renderer: Renderer2) {}
+  private highlightTagElements: Array<{
+    element: HTMLElement;
+    clientRect: ClientRect;
+  }>;
+
+  private mouseHoveredTag: TagMouseEvent | undefined;
+
+  constructor(private renderer: Renderer2, private cdr: ChangeDetectorRef) {}
 
   /**
    * @private
@@ -148,21 +188,60 @@ export class TextInputHighlightComponent implements OnChanges, OnDestroy {
         }),
         this.renderer.listen(this.textInputElement, 'scroll', () => {
           this.highlightElement.nativeElement.scrollTop = this.textInputElement.scrollTop;
+          this.highlightTagElements = this.highlightTagElements.map(tag => {
+            tag.clientRect = tag.element.getBoundingClientRect();
+            return tag;
+          });
         })
       ];
+
+      // only add event listeners if the host component actually asks for it
+      if (this.tagClick.observers.length > 0) {
+        const onClick = this.renderer.listen(
+          this.textInputElement,
+          'click',
+          event => {
+            this.handleTextareaMouseEvent(event, 'click');
+          }
+        );
+        this.textareaEventListeners.push(onClick);
+      }
+
+      if (this.tagMouseEnter.observers.length > 0) {
+        const onMouseMove = this.renderer.listen(
+          this.textInputElement,
+          'mousemove',
+          event => {
+            this.handleTextareaMouseEvent(event, 'mousemove');
+          }
+        );
+        const onMouseLeave = this.renderer.listen(
+          this.textInputElement,
+          'mouseleave',
+          event => {
+            if (this.mouseHoveredTag) {
+              this.tagMouseLeave.emit(this.mouseHoveredTag);
+              this.mouseHoveredTag = undefined;
+            }
+          }
+        );
+        this.textareaEventListeners.push(onMouseMove);
+        this.textareaEventListeners.push(onMouseLeave);
+      }
 
       this.addTags();
     });
   }
 
   private addTags() {
-    let textareaValue = this.textInputElement.value;
+    const textareaValue = this.textInputElement.value;
 
     const prevTags: HighlightTag[] = [];
+    const parts: string[] = [];
 
     [...this.tags]
       .sort((tagA, tagB) => {
-        return tagB.indices.start - tagA.indices.start;
+        return tagA.indices.start - tagB.indices.start;
       })
       .forEach(tag => {
         if (tag.indices.start > tag.indices.end) {
@@ -182,9 +261,7 @@ export class TextInputHighlightComponent implements OnChanges, OnDestroy {
           }
         });
 
-        // TODO - implement this as an ngFor of items that is generated in the template
-        const before = textareaValue.slice(0, tag.indices.start);
-        const after = textareaValue.slice(tag.indices.end);
+        // TODO - implement this as an ngFor of items that is generated in the template for a cleaner solution
 
         const expectedTagLength = tag.indices.end - tag.indices.start;
         const tagContents = textareaValue.slice(
@@ -192,16 +269,61 @@ export class TextInputHighlightComponent implements OnChanges, OnDestroy {
           tag.indices.end
         );
         if (tagContents.length === expectedTagLength) {
+          const previousIndex =
+            prevTags.length > 0 ? prevTags[prevTags.length - 1].indices.end : 0;
+          const before = textareaValue.slice(previousIndex, tag.indices.start);
+          parts.push(escapeHtml(before));
           const cssClass = tag.cssClass || this.tagCssClass;
-          textareaValue =
-            before +
-            `<span class="text-highlight-tag${cssClass
-              ? ' '
-              : ''}${cssClass}">${tagContents}</span>` +
-            after;
+          const tagId = tagIndexIdPrefix + this.tags.indexOf(tag);
+          // text-highlight-tag-id-${id} is used instead of a data attribute to prevent an angular sanitization warning
+          parts.push(
+            `<span class="text-highlight-tag ${tagId} ${cssClass}">${escapeHtml(
+              tagContents
+            )}</span>`
+          );
           prevTags.push(tag);
         }
       });
-    this.highlightedText = `${textareaValue}&nbsp;`;
+    const remainingIndex =
+      prevTags.length > 0 ? prevTags[prevTags.length - 1].indices.end : 0;
+    const remaining = textareaValue.slice(remainingIndex);
+    parts.push(escapeHtml(remaining));
+    parts.push('&nbsp;');
+    this.highlightedText = parts.join('');
+    this.cdr.detectChanges();
+    this.highlightTagElements = Array.from(
+      this.highlightElement.nativeElement.getElementsByTagName('span')
+    ).map((element: HTMLElement) => {
+      return { element, clientRect: element.getBoundingClientRect() };
+    });
+  }
+
+  private handleTextareaMouseEvent(
+    event: MouseEvent,
+    eventName: 'click' | 'mousemove'
+  ) {
+    const matchingTagIndex = this.highlightTagElements.findIndex(elm =>
+      isCoordinateWithinRect(elm.clientRect, event.clientX, event.clientY)
+    );
+    if (matchingTagIndex > -1) {
+      const target = this.highlightTagElements[matchingTagIndex].element;
+      const tagClass = Array.from(target.classList).find(className =>
+        className.startsWith(tagIndexIdPrefix)
+      );
+      if (tagClass) {
+        const tagId = tagClass.replace(tagIndexIdPrefix, '');
+        const tag: HighlightTag = this.tags[+tagId];
+        const tagMouseEvent = { tag, target };
+        if (eventName === 'click') {
+          this.tagClick.emit(tagMouseEvent);
+        } else if (!this.mouseHoveredTag) {
+          this.mouseHoveredTag = tagMouseEvent;
+          this.tagMouseEnter.emit(tagMouseEvent);
+        }
+      }
+    } else if (eventName === 'mousemove' && this.mouseHoveredTag) {
+      this.tagMouseLeave.emit(this.mouseHoveredTag);
+      this.mouseHoveredTag = undefined;
+    }
   }
 }